Run test_functional_api.py with both legacy and native funcol impls
Additional changes: tests in test_functional_api.py uses multi-threaded pg which is implemented in Python. For the native ops to call into the Python pg implementation, glue code in PyProcessGroup is required for each collective. This PR also adds a few pieces of previously missing glue code, which are necessary for running test_functional_api.py with native funcol.

ghstack-source-id: 0c5d4b7b0f3f246b8a1b372ac6eab55a8d75c36c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/119982